+++ packages/realm-react/CHANGELOG.md
@@ -27,7 +27,7 @@ const MyApp() = () => {
 import { RealmProvider, RealmProviderFallback } from "@realm/react";
 
 const Fallback: RealmProviderFallback = ({ progress }) => {
-	return <Text>Loading: {progress}/1.0</Text>;
+	return <Text>Loading:{(100 * progress).toFixed()}%</Text>;
 }
 
 const MyApp() = () => {