+++ packages/realm-react/CHANGELOG.md
@@ -1,10 +1,12 @@
 ## vNext (TBD)
 
+### Note: Compatibility with Realm has been bumped to [Realm 12.12.0](https://github.com/realm/realm-js/releases/tag/v12.12.0) to add better support for progress notifications.
+
 ### Deprecations
 * None
 
 ### Enhancements
-* Added `useProgress` hook which provides a convenient way to access Realm's progress information. It works in a similar way as `realm.addProgressNotification`. **Note: this feature is only intended for flexible sync with the new estimate-based progress notifications.** ([#6797](https://github.com/realm/realm-js/issues/6797))
+* Added `useProgress` hook which provides a convenient way to access Realm's progress information. It works in a similar way as `realm.addProgressNotification`. **Note: this feature is using the new estimate-based progress notifications introduced in [Realm 12.12.0](https://github.com/realm/realm-js/releases/tag/v12.12.0).** ([#6797](https://github.com/realm/realm-js/issues/6797))
 ```tsx
 import { RealmProvider, ProgressDirection, ProgressMode } from "@realm/react";